+++ EduOS.Persistence/Repositories/EnrollmentRepository.cs
@@ -11,8 +11,8 @@ namespace EduOS.Persistence.Repositories
 
         public async Task<List<Enrollment>> GetByStudentIdAsync(long studentId) => await _dbSet.Include(e => e.Class).Include(e => e.Section).Include(e => e.AcademicYear).Where(e => e.StudentId == studentId).OrderByDescending(e => e.EnrollmentDate).ToListAsync();
 
-        public async Task<Enrollment?> GetCurrentAsync(long studentId, int academicYearId) => await _dbSet.FirstOrDefaultAsync(e => e.StudentId == studentId && e.AcademicYearId == academicYearId && e.IsActive);
+        public async Task<Enrollment?> GetCurrentAsync(long studentId, long academicYearId) => await _dbSet.FirstOrDefaultAsync(e => e.StudentId == studentId && e.AcademicYearId == academicYearId && e.IsActive);
 
-        public async Task<List<Enrollment>> GetByClassSectionAsync(int classId, int sectionId, int academicYearId) => await _dbSet.Include(e => e.Student).Where(e => e.ClassId == classId && e.SectionId == sectionId && e.AcademicYearId == academicYearId && e.IsActive).OrderBy(e => e.Roll).ToListAsync();
+        public async Task<List<Enrollment>> GetByClassSectionAsync(long classId, long sectionId, long academicYearId) => await _dbSet.Include(e => e.Student).Where(e => e.ClassId == classId && e.SectionId == sectionId && e.AcademicYearId == academicYearId && e.IsActive).OrderBy(e => e.Roll).ToListAsync();
     }
 }
